+++ packages/api/src/index.ts
@@ -10,5 +10,6 @@ export * from "./data/fuelLogs";
 export * from "./data/invites";
 export * from "./data/mapReports";
 export * from "./data/memberships";
+export * from "./data/notifications";
 export * from "./data/tripEvents";
 export * from "./data/vehicles";